+++ app/src/main/java/com/kabos/spotifydj/di/OtherModule.kt
@@ -1,0 +1,23 @@
+package com.kabos.spotifydj.di
+
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.converter.moshi.MoshiConverterFactory
+import javax.inject.Singleton
+
+@InstallIn(SingletonComponent::class)
+@Module
+object OtherModule {
+    @Singleton
+    @Provides
+    fun provideMoshiConverterFactory(): MoshiConverterFactory {
+        val moshi = Moshi.Builder()
+            .add(KotlinJsonAdapterFactory())
+            .build()
+        return MoshiConverterFactory.create(moshi)
+    }
+}
