+++ app/src/main/java/com/kabos/spotifydj/di/RetrofitModule.kt
@@ -1,23 +1,22 @@
 package com.kabos.spotifydj.di
 
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import com.kabos.spotifydj.model.apiConstants.ApiConstants
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
-object MoshiConverterFactoryModule {
+object RetrofitModule {
     @Singleton
     @Provides
-    fun provideMoshiConverterFactory(): MoshiConverterFactory {
-        val moshi = Moshi.Builder()
-            .add(KotlinJsonAdapterFactory())
+    fun provideSpotifyApi(moshiConverterFactory: MoshiConverterFactory): Retrofit =
+        Retrofit.Builder()
+            .baseUrl(ApiConstants.ENDPOINT)
+            .addConverterFactory(moshiConverterFactory)
             .build()
-        return MoshiConverterFactory.create(moshi)
-    }
 }
