+++ app/src/main/java/com/kabos/spotifydj/ui/activity/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.kabos.spotifydj.ui
+package com.kabos.spotifydj.ui.activity
 
 import android.content.ComponentName
 import android.content.Intent
@@ -13,10 +13,10 @@ import androidx.lifecycle.LiveData
 import com.kabos.spotifydj.R
 import com.kabos.spotifydj.databinding.ActivityMainBinding
 import com.kabos.spotifydj.util.OneShotEvent
-import com.kabos.spotifydj.viewModel.PlaylistViewModel
-import com.kabos.spotifydj.viewModel.RecommendViewModel
-import com.kabos.spotifydj.viewModel.SearchViewModel
-import com.kabos.spotifydj.viewModel.UserViewModel
+import com.kabos.spotifydj.ui.viewModel.PlaylistViewModel
+import com.kabos.spotifydj.ui.viewModel.RecommendViewModel
+import com.kabos.spotifydj.ui.viewModel.SearchViewModel
+import com.kabos.spotifydj.ui.viewModel.UserViewModel
 import com.spotify.sdk.android.auth.AuthorizationClient
 import com.spotify.sdk.android.auth.AuthorizationRequest
 import com.spotify.sdk.android.auth.AuthorizationResponse
